[774] Move operation options into separate file
