+++ bob-backend/src/lib.rs
@@ -22,7 +22,8 @@ pub(crate) mod prelude {
             cluster::Cluster as ClusterConfig,
             node::{BackendType, Node as NodeConfig, Pearl as PearlConfig},
         },
-        data::{BobData, BobKey, BobMeta, BobPutOptions, BobGetOptions, BobDeleteOptions, DiskPath, VDiskId},
+        data::{BobData, BobKey, BobMeta, DiskPath, VDiskId},
+        operation_options::{BobPutOptions, BobGetOptions, BobDeleteOptions},
         node::NodeName,
         error::Error,
         mapper::Virtual,
