+++ bob-common/src/lib.rs
@@ -2,6 +2,7 @@ pub mod bob_client;
 pub mod configs;
 mod name_types; // Private module. Inner types should be re-exported where they are needed
 pub mod data;
+pub mod operation_options;
 pub mod error;
 pub mod mapper;
 pub mod metrics;
