+++ bob/src/lib.rs
@@ -50,7 +50,8 @@ mod prelude {
     pub use bob_common::{
         bob_client::{BobClient, Factory},
         configs::node::Node as NodeConfig,
-        data::{BobData, BobKey, BobMeta, BobPutOptions, BobGetOptions, BobDeleteOptions, DiskPath, VDiskId},
+        data::{BobData, BobKey, BobMeta, DiskPath, VDiskId},
+        operation_options::{BobPutOptions, BobGetOptions, BobDeleteOptions},
         error::Error,
         mapper::Virtual,
         metrics::{
